requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==4.2.2
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==4.0.2
attrdict==2.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.6.2
bpytop==1.0.68
Brotli==1.0.9
cachetools==5.3.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cloud-init==23.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
deskew==1.4.3
distro==1.7.0
distro-info==1.1+ubuntu0.1
docopt==0.6.2
entrypoints==0.4
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.29.1
frozenlist==1.3.3
fs==2.4.12
fuzzywuzzy==0.18.0
gast==0.5.2
gitdb==4.0.10
GitPython==3.1.30
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio==2.28.1
importlib-metadata==4.6.4
imutils==0.5.4
incremental==21.3.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
lxml==4.8.0
lz4==3.1.3+dfsg
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.1
notebook==6.4.8
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.1
pdf2image==1.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.13
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
py==1.10.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydeck==0.8.0
Pygments==2.14.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylatexenc==2.10
Pympler==1.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
PyPDF3==1.0.6
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pythran==0.10.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.25.1
rfc3986==1.5.0
rich==13.3.1
scikit-image==0.20.0
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semver==2.13.0
Send2Trash==1.8.1b0
service-identity==18.1.0
six==1.16.0
smmap==5.0.0
sos==4.5.6
soupsieve==2.3.1
ssh-import-id==5.11
streamlit==1.18.1
streamlit-launchpad==0.0.11
sympy==1.9
systemd-python==234
terminado==0.13.1
testpath==0.5.0
tifffile==2023.4.12
tk==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.65.0
traitlets==5.1.1
trash-cli==0.17.1.14
twilio==8.2.0
Twisted==22.1.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
validators==0.20.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
Werkzeug==2.3.4
widgetsnbextension==2.0.0
xkit==0.0.0
yarg==0.1.9
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
